ns: rewrite IReadOnlyApplicationRecordInterface
